Fix : 로비 화면 상점 구현 - 도장 수정
로비 화면에서 구매하기와 버튼 관련 스크립트가 한 스크립트에 작성되어 있어 분리 작업 - SampleScene
