+++ TeamODD.ver0.0.3/Assets/Room/StoreLobbyMovement.cs.meta
@@ -1,0 +1,15 @@
+fileFormatVersion: 2
+guid: ca0e01c44f98ac645bc9b22eb20f60a8
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences:
+  - STGroup: {instanceID: 0}
+  - StempsImage: {instanceID: 0}
+  - ShapeImage: {instanceID: 0}
+  - Testsss: {fileID: 21300000, guid: 0a89e501353a7ea48bf9a70b55a2c5c8, type: 3}
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
