+++ TeamODD.ver0.0.3/Assets/Scenes/SampleScene.unity
@@ -430,8 +430,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 891567109}
-        m_TargetAssemblyTypeName: 
+      - m_Target: {fileID: 891567110}
+        m_TargetAssemblyTypeName: StoreLobbyMovement, Assembly-CSharp
         m_MethodName: FAR
         m_Mode: 1
         m_Arguments:
@@ -1100,6 +1100,10 @@ MonoBehaviour:
   Ink_1_: {fileID: 21300000, guid: 3c29a6114368b5d44bc7417b152b2962, type: 3}
   Ink_2_: {fileID: 0}
   Ink_3_: {fileID: 0}
+  Close_Ink_N_: {fileID: 0}
+  Close_Ink_1_: {fileID: 0}
+  Close_Ink_2_: {fileID: 0}
+  Close_Ink_3_: {fileID: 0}
   ERROR_: {fileID: 0}
 --- !u!1 &566586749
 GameObject:
@@ -1417,11 +1421,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   Canvas: {fileID: 0}
-  STGroup: {fileID: 0}
-  StempsImage: {fileID: 619711843}
-  ShapeImage: {fileID: 1683917240}
   MatchLoading: {fileID: 0}
-  Testsss: {fileID: 0}
 --- !u!1 &807496975
 GameObject:
   m_ObjectHideFlags: 0
@@ -1509,6 +1509,7 @@ GameObject:
   - component: {fileID: 891567108}
   - component: {fileID: 891567107}
   - component: {fileID: 891567109}
+  - component: {fileID: 891567110}
   m_Layer: 5
   m_Name: GroupGameButton
   m_TagString: Untagged
@@ -1596,10 +1597,22 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   Canvas: {fileID: 0}
+  MatchLoading: {fileID: 0}
+--- !u!114 &891567110
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 891567105}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ca0e01c44f98ac645bc9b22eb20f60a8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
   STGroup: {fileID: 1676324739}
   StempsImage: {fileID: 619711843}
   ShapeImage: {fileID: 1683917240}
-  MatchLoading: {fileID: 0}
   Testsss: {fileID: 21300000, guid: 0a89e501353a7ea48bf9a70b55a2c5c8, type: 3}
 --- !u!1 &893383195
 GameObject:
@@ -2578,6 +2591,7 @@ GameObject:
   - component: {fileID: 1961230560}
   - component: {fileID: 1961230564}
   - component: {fileID: 1961230565}
+  - component: {fileID: 1961230566}
   m_Layer: 5
   m_Name: Canv_Hub
   m_TagString: Untagged
@@ -2693,6 +2707,22 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 749e50af13f148347acde49f853c0c49, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &1961230566
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1961230559}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: ca0e01c44f98ac645bc9b22eb20f60a8, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  STGroup: {fileID: 1676324739}
+  StempsImage: {fileID: 619711843}
+  ShapeImage: {fileID: 1683917240}
+  Testsss: {fileID: 21300000, guid: 0a89e501353a7ea48bf9a70b55a2c5c8, type: 3}
 --- !u!1 &1978435451
 GameObject:
   m_ObjectHideFlags: 0
@@ -3118,8 +3148,8 @@ MonoBehaviour:
   m_OnClick:
     m_PersistentCalls:
       m_Calls:
-      - m_Target: {fileID: 891567109}
-        m_TargetAssemblyTypeName: 
+      - m_Target: {fileID: 891567110}
+        m_TargetAssemblyTypeName: StoreLobbyMovement, Assembly-CSharp
         m_MethodName: BAR
         m_Mode: 1
         m_Arguments:
